Rename public API for room lifecycle and status
We:

- rename the RoomLifecycle enum to RoomLifecycleStatus (which makes more
  sense);
- rename the RoomStatus protocol to RoomLifecycle and rename its
  `current` property to `status` ("current" what?; it wasn’t clear, and
  was inconsistent with the core SDKs, which only use `current` for state
  _changes_)

The net effect is that instead of writing `room.status.current` you now
write `room.lifecycle.status`, which still seems pretty readable and
gives the benefits listed above.

This change has been agreed with the Chat team and the decision is
recorded in [1]; they will make this change in JS.

As part of this, I’ve also corrected references to room “state” to the
correct terminology of “status” (in [2] Andy explains why they used the
word “status”, choosing to reserve “state” for some larger composite
state).

Outstanding spec question about whether to rename the RoomInFailedState
error [3]; that can wait.

[1] https://ably.atlassian.net/wiki/spaces/CHA/pages/3307405320/SDK+Divergence+Decision+Log
[2] https://github.com/ably-labs/ably-chat-swift/issues/73#issuecomment-2385770039
[3] https://github.com/ably/specification/pull/200/files#r1783542331