Expose lifecycle operation waits
Allow tests to find out that one lifecycle operation is waiting for
another. This will be needed for testing spec point CHA-RL1d, which
states that an ATTACH operation should wait for an in-progress operation
to complete before proceeding; I couldn’t think of any way to test this
(e.g. looking for some side effect) with the information that we
currently have, hence this testing API.