Split room lifecycle “attaching” status in two
This will be used for implementing CHA-RL4b7 (from the spec referenced
in 20f21c7), which allows a contributor state change to put the room
into an ATTACHING status outside of the context of a room lifecycle
operation.

In this situation, it may also wish to attach an error to the ATTACHED
room status, so update the public API to reflect this.