Implement “previously attached” part of CHA-RL4b1
Based on [1] at d9af7bf. I’m assuming that the “no call” in this spec
point is a typo and means “a call”; have asked in [2].

Needed for when we integrate the room lifecycle manager into the rest of
the SDK in #47 (else there will be a crash when running the integration
tests, because a channel’s first ATTACHED status change has resumed ==
false and our precondition will fail).

[1] https://github.com/ably/specification/pull/200
[2] https://github.com/ably/specification/pull/200#discussion_r1831431062