Make room responsible for channel fetching
The room needs to have access to its contributing channels so that, when
we integrate the RoomLifecycleManager into the SDK in #47, it can use
them to create a RoomLifecycleManager instance. So, let’s make the room
fetch the channels and then pass them to the features. This way, we only
fetch each channel once.

Also, in an upcoming commit it will be the room who releases the
channels when `release()` is called on it, giving another reason why it
makes sense for the room to fetch the channels, I think.