Comment out code for setting presence modes
This code is not actually doing anything at the moment; the way in which
we’re currently fetching channels means that these modes get
overwritten, which we’ll fix in #132. But, when we fix the
aforementioned, we’ll see that setting these flags leads to some weird
Realtime behaviours, so we’ll turn them off for now and restore them
separately in #133.