Implement CHA-RL5a1
This spec point requires us to implement a special behaviour to handle
the fact that multiple contributors can share a channel. I have decided,
instead, to make it so that each channel has precisely one lifecycle
contributor. I think this is a simpler, functionally equivalent approach
and have suggested it in [1].

[1] https://github.com/ably/specification/issues/240