Introduce a DiscontinuityEvent type
This is what you’ll now receive when you subscribe to a room feature’s
discontinuities. I should have done this in 30cfd30, when we started
allowing a discontinuity’s error to be nil; it’s not nice to work with
an AsyncSequence that emits optional values, because it makes the return
value of operators like `first` ambiguous; did you reach the end of the
sequence or did you get a discontinuity without an associated error?