Fix agent string
We don’t have a proper solution for analytics yet ([1] is on hold and
has been removed from mobile beta scope), but in a conversation in the
team we decided that we might as well at least keep the `agent` channel
parameter, which will give us information about channel attaches. So,
name it in line with the JS SDK, which uses "chat-js"; i.e. use the name
of the programming language ("ios" is misleading since we might be
running on macOS or tvOS).

[1] https://github.com/ably/specification/issues/201