Add an “all features enabled” room option
In 20e7f5f I said

> RoomOptionsDefaults in JS is instead implemented here by giving the
> *Options types a no-args initializer that populates the default values.

but I had not noticed that JS’s RoomOptionsDefaults is not actually a
“default”, since the JS SDK does not even have a concept of a “default”
RoomOptions. What it actually is is a RoomOptions in which all of the
room features are switched on; we discussed this today in standup and
Andy said that the intention of this value was to give users who are
just playing around with the SDK an easy way to turn all features on.
So, here, I add such an API for Swift, but with a name that more
accurately describes its intention. Andy’s created [1] to revisit the JS
naming.

[1] https://ably.atlassian.net/browse/CHA-766