Rename `Message.latestAction` to `action`
Upon our integration tests starting to fail, we noticed that Realtime
seems to have renamed the `latestAction` property to `action`. This
change has not yet been reflected in the spec, but has been in JS [1].

I’ve created an issue to get the spec updated [2] but since this test
failure is stopping us from merging anything else I am going to rename
this property until we get a more definitive answer.

Part of #169.

[1] https://github.com/ably/ably-chat-js/pull/427
[2] https://github.com/ably/specification/issues/254