Make bufferingPolicy parameter optional in public API
When I introduced the public API in 20e7f5f, I didn’t want users to be
in a situation where there was an buffer of unbounded size growing
without them even knowing about it, and I hence forced them to be
explicit about their choice of buffering policy.

But, a couple of things:

- Marat pointed out that this makes our code examples and README look
  quite cluttered, with `bufferingPolicy` arguments all over the place.
  This distracts from demonstrating the functionality of the library.

- Swift’s own `AsyncStream` defaults to a buffering policy of
  `.unbounded`. If doing so is good enough for them, then it’s probably
  good enough for us.

So, default to `.unbounded`.

Resolves #172.