Rename subscribeToDiscontinuities to onDiscontinuity
I remember that when I wrote the public API in 20e7f5f, I was a bit
reluctant to copy JS’s `on*` naming, because that read to me more like
an API that was designed for receiving a listener (i.e. “_on_ this
event, call this listener”). But, I think that after some back and forth
with Andy, we decided to stick to the `on*` naming, hence
`onStatusChange`. For some reason, though, I didn’t apply this decision
to `subscribeToDiscontinuties`. So, do that.

Resolves #173.