wip
improvements:

- supports bool, nested objects etc
- your presence data can be anything JSONValue, not just an object (for
  now, this makes the call site a bit messier but see below we'll fix
that)
- distinguishing between not passing an argument and not
- you don't have to pass userCustomData explicitly

TODO explain the JS behaviour this is copying, and create a spec issue
for me to refer to

TODO have to decide the right thing to do for the arg-having variants;
should they take nil or no?

fluent api to come in separate commit

i am not allowing the passing of PresenceData? because it increases the
confusion of the meaning of `nil`. but for now i have kept it in the
`PresenceMember` thing, `nil` meaning not passed