Improve API and internals for presence data
Public API improvements:

- You can now pass any JSON value as presence data (previously, the
  top-level value had to be an object, and you could not pass arrays or
  nested objects). This brings us in line with CHA-PR2a.

- Do not expose the `userCustomData` property of the presence data
  object in the public API; it’s an implementation detail.

- Conform to the `ExpressibleBy*Literal` protocols, making it easier to
  create a JSON value.

I have deliberately chosen to make the data-arg variants of the presence
operation methods take a non-optional PresenceData. This is to minimise
confusion between the absence of presence data and a presence data with
JSON value `null`. This is how I wrote this API in 20e7f5f, but I didn’t
restore it properly in 4ee16bd.

The JSONValue type introduced here is based on the example given in [1].

Internals improvements:

- Fix the presence object that gets passed to ably-cocoa when the user
  specifies no presence data; we were previously passing an empty string
  as the presence data in this case. (See below for where I got the
  “correct” behaviour from.)

- Simplify the way in which we decode the presence data received from
  ably-cocoa (i.e. don’t do a round-trip of JSON serialization and
  deserialization); this comes at the cost of not getting a little bit for
  free from Swift’s serialization mechanism, but I think it’s worth it

The behaviour of how to map the chat presence data public API to the
object exchanged with the core SDK is not currently fully specified. So,
the behaviour that I’ve implemented here is based on the behaviour of
the JS Chat SDK at 69ea478. I’ve created spec issue [2] in order to
specify this stuff properly, but I’m in a slight rush to get this public
API fixed before we release our first beta, so I’ll address this later.

Resolves #178.

[1] https://www.douggregor.net/posts/swift-for-cxx-practitioners-literals/
[2] https://github.com/ably/specification/issues/256