Remove usage of Decodable
Switch to using JSONDecodable internally. This allows us to avoid the
dance of encoding an ably-cocoa object to a Data just so that we can
decode it using JSONDecoder. It also gives us consistency in how we
handle data objects received from ably-cocoa.

This means that we now have to write a bit more code in order to decode
objects manually, but I think we can live with it. (It might also be
that we can revisit in the future and either automatically generate the
JSONDecodable conformance using macros, or that we can find some other,
better way of bridging with the compiler’s synthesised Decodable
conformance.)

Resolves #84.