Fix sending and receiving of reaction metadata and headers
Similar issue to the one which was fixed in 7fcab5c.

Haven’t added tests for the handling of the incoming message, since
there were no existing tests to update and I didn’t feel like writing
them; will leave this for whoever writes these tests in #88.

The approach that I’ve taken here of using a DTO is consistent with the
approach that we use for presence data. I should have done this in
7fcab5c too; will do it separately.

The spec doesn’t describe how to behave if the user doesn’t pass headers
or metadata; I’ve taken the behaviour of passing an empty object from
the JS Chat SDK at 6d1b04a. Have created spec issue [1] for specifying
this.

Resolves #198.

[1] https://github.com/ably/specification/issues/256