Add a dictionary specialisation of JSONCodable
Add a JSONObjectCodable protocol which represents a type that can be
serialised to or deserialised from a [String : JSONValue] dictionary.
As well as slightly simplifying the deserialisation of a type (you don't
have to do the "is it an object" check yourself), this will be necessary
when, in an upcoming commit, we try to pass a serialised value to the
`extras` parameter of ably-cocoa’s publish, which expects an
`ARTJsonCompatible` value.