Change the type of LogContext
We don’t actually emit any context at the moment, I believe, nor do we
tell users what they’re meant to do with it. But, still, before our
first release of the public API, let’s use a type that the user can at
least easily pass through to a logging backend.

We still have #8 for figuring out what this context is actually meant to
be used for.