Throw ARTErrorInfo from public API
This updates our public API to use typed throws, and specifically to
always throw ARTErrorInfo. This makes us consistent with ably-cocoa, and
allows users to access the error's `statusCode` to make decisions about
whether they can retry the action that caused the error.

I've gone for a light-touch approach of not changing any of the various
error types that exist inside the SDK; inside I've hidden them all
behind a common InternalError type, which can be converted to an
ARTErrorInfo. We can revisit this in #32 (i.e. should we keep this rich
internal Swift error types or just throw ARTErrorInfo internally?)

Resolves #43.