Move `realtime.request` usage into an async helper
Our current usage of `realtime.request` is quite messy at the call site,
and will get more so when we need to use typed throws throughout the
codebase for #43. So move it to a helper (I've changed it to use typed
throws whilst I'm at it).

This also opens the door to us using a
RoomLifecycleContributorChannel-like protocol internally when making
requests, which will simplify mocking.