Use a single channel for all features
Resolves #242.

The API changes are based on the JS API changes from [1] at acd0a36.
The behavioural changes are based on [2] at 3cc7e2f.

This commit is quite large, and combines API and behavioural changes,
because the new specification points are written with reference to the
new API and so it was hard to disentangle the two.

Key changes to public API:

- Updated room options API:
  - there's no concept of a feature being turned "on" any more, and
    correspondingly no error when you try and use a feature that's not
    turned on

  - removed `allFeaturesEnabled`, since the equivalent has been removed
    from JS

  - you can now fetch a room without passing a room options; this uses
    the default options

- Removed feature-specific error codes

- Discontinuities now always have an error

Key changes to interactions with Realtime:

- Switched to v3 of the REST API

- Use a single channel for all room features; this simplifies the room
  lifecycle manager and the DefaultRoom class

- Room reactions now use ephemeral messages

- Resolves #133 (setting of presence-related channel modes).

Internal changes:

- Removed FeatureChannel and RoomLifecycleContributor

[1] https://github.com/ably/ably-chat-js/pull/505
[2] https://github.com/ably/specification/pull/282