wip switch to using the async interface internally
the aim is to end up not having multiple interfaces, and to be able to
write mocks more easily

wip

further

wip

further

ok, for now we won't try and replace any of the existing methods, we can
try and make them Swiftier as need be (i.e. in a later commit, when we
need to be able to mock them)

further

TODO make sure that we're not accidentally using the underlying via `channel` property

further

further

further

further

yeah the "don't accidentally call methods on `channel` is a bit of a
problem", especially if we remove the `Async` from the method names,
because we might accidentally instead be calling the no-callback
non-async version.

further

TODO remember to go and make the API more friendly afterwards

further

further

further

further

further

not sure why I switched the weak from channel to self, but feature
channel is not class-bound

working on fixing tests

further

wip

further

wip checking contenst of channel mock

further

further

further

further

further

further