Don't use AttachmentFailed in DefaultMessages
We'll be getting rid of the *AttachmentFailed error codes as part of the
upcoming single-channel migration. The spec does not say to use the
MessagesAttachmentFailed error code as part of the messages feature, so
remove its use here and replace it with something generic. (I don't know
what _is_ the right thing to use here since I'm not familiar with this
functionality; from a glance of the spec it doesn't seem like the code
we've written here comes from any behaviour described in the spec.)