Convert public API and internals to @MainActor
In 7d6acde we made the decision to use `actor` types to manage the
mutable state of the SDK. Whilst we've been able to build the SDK using
this approach, it has proved cumbersome.

The public API is littered with `async` methods and getters for things
that really should be synchronous, such as adding a subscription or
fetching a room's state. This deviates from the JavaScript API and makes
call sites look complicated.

The use of `async` for fetching room state also means that it's
impossible for a user to, for example, check a room's state and then
reason about what operations they should be able to attempt to perform
on the room (i.e. without the SDK immediately rejecting the attempt),
because the room state can change under their feet. This is one of a
class of threading-related problems that I describe in #260.

Internally, we find ourselves having to write `await` to do basic things
like fetching the room lifecycle manager's state or to find out the
arguments with which a mock method was called. And similarly, one
internal component can't rely on the state that's isolated to a
different internal isolation domain in order to make a decision, because
this state might change under its feet.

To address this, here I have decided to isolate all of the SDK's mutable
state to the main actor. This simplifies the experience of developing
the SDK, and it simplifies the experience of interacting with it from
the main thread, which is where most of our users interact with it from.

(Note that this does not prevent us from using background tasks
internally for moving heavy processing off the main thread; I'm
specifically talking about _state_.)

This also moves us closer to being able to address #49 (keeping our
internal state in sync with that of the underlying ably-cocoa objects).

Note that this change does decrease the parallelisability of our test
suite. But it still runs very fast.

Have removed:

- all `actor` types (both in SDK and mocks)
- nearly all locking (there are a couple of non-MainActor types for
  which we still have it)

There might be some simplifications that can be made for the lifecycle
manager tests now, but I'm not going to look at them now because that
class will change a lot for the single-channel work in #242.

I'll update the ably.com documentation in #254.

Even after this change, Messages.subscribe is still `async` because its
implementation requires it to be. I believe that this is a bug; have
created #257 for it.

Resolves #232.