Temporarily remove typed throws from PaginatedResult
With this change, the SDK now compiles in Xcode 16.3 without crashing
the compiler.

This is a breaking public API change but we're not at 1.x yet and also I
hope that we can revert it soon, once the referenced compiler crash is
fixed. If it turns out that the compiler crash isn't going to be fixed
any time soon, then we might need to try and find a workaround that
preserves the typed throws (I've already put a bit of time into trying
to find one, but didn't succeed and thought it wasn't worth spending
more time on yet).

Resolves #267.