Rename channel mock's emitStateChange
I want us to start referring to a generic ARTChannelStateChange object
as a "channel event", not a "channel state change", since, despite its
unfortunate naming, such an object actually refers either a state change
or an UPDATE event. There are still plenty of places in the codebase
where the naming needs to be updated, which I'll address during the
upcoming single-channel work to avoid churn.