Rename PresenceQuery to PresenceParams
An API of get(params _: PresenceQuery) doesn't read very well, so we
either need to change the type name or the parameter name. I've decided
to go for consistency with the chat-js API instead of consistency with
ably-cocoa (even though I prefer the ably-cocoa name).

(The underlying issue here is that ably-cocoa and ably-js use different
names for the type that describes a presence query.)