Keep the PresenceParams type
I originally created this as a way of getting around some compiler
concurrency checking warnings when we used ably-cocoa's
ARTRealtimePresenceQuery type. Now that we've switched to @MainActor in
fc83fc1, we could switch back to using the ably-cocoa type.

However, I think it's best that we keep this Swift type, because I think
it makes more sense to use a value type instead of the mutable reference
type used by ably-cocoa.

Documentation is adapted from core SDK canonical docstrings in spec repo.

Resolves #64.