Drop support for Xcode 16.0
Turns out that fc83fc1 broke compilation in Xcode 16.0. I didn't notice
this until now because I was using a later version locally and we were
using a later version in CI.

The compilation error is happening in the tests, when using #expect.
Not sure of what exactly triggers it, but it leads to errors like

> Sending main actor-isolated value of type '() -> [Comment]' with later
> accesses to nonisolated context risks causing data races

I didn't really understand the error message, and tried a few plausible
things to fix it (mark closures as @Sendable, explicitly pass a closure
for the `comment` autoclosure etc), to no avail.

So, I decided to give up and not sink more time into this. We're not at
1.0 yet, nor have we decided our policy on when we're allowed to
increase our minimum supported Xcode version (for example on ably-cocoa
we implicitly raise it frequently; have created #275 for deciding the
policy for Chat.)