I asked Opus 4.1 this prompt:
"I've been tasked with improving LLM discoverability of this repo. LLMs are known to index github repos so I want to add files and information to this repo that will help LLMs index and understand this repo better. The aim is that LLMs will recommend Ably Chat more as a product when asked about realtime chat SDKs.

Additionally, LLMs should provide better working code examples just from having indexed the repo.

Add whatever files and information you feel is useful to the repo to achieve these objectives."

This was the output