Preserve type information of core SDK objects
Change the ChatClient.realtime and Room.channel properties of
DefaultChatClient to return ARTRealtime and ARTRealtimeChannel, instead
of flattening them down to existential types.

In order to achieve this without having to make DefaultChatClient
generic over the client, we now only allow a DefaultChatClient to be
created with a concrete ARTRealtime client. There is little downside to
this; users have no need to use a mock Realtime client, and for nearly
all of our unit testing we just directly instantiate the internal
components with mocks. (None of this prevents users from creating their
own mock ChatClient, which as mentioned in CONTRIBUTING.md remains an
explicit goal of the library.)

(Note we could have put a primary associated type on ChatClient or Room,
too, but for now I've just kept one where it's needed; on Rooms to
enable the `some Room<ARTRealtimeChannel>` type, which lets us keep
DefaultRooms private. I believe that we can add further primary
associated types in a non-breaking fashion later, if we see a need for
it.)

Note that there's a behavioural change here — Room.channel now returns
the ARTRealtimeChannel instead of the ARTWrapperSDKProxyRealtimeChannel,
meaning that REST operations initiated via this channel are no longer
attributed to the chat SDK. This is the correct behaviour per the ADR
referenced in 67b146a.