Remove Protocol from public protocol names
This makes us consistent with the rest of the public protocol naming.

A note re the pros and cons of having "Protocol" in protocol names:

- Pros: It makes them easier to use when declaring an associated type or
  typealias that shares a name (e.g. `associatedtype Connection:
  ConnectionProtocol` instead of `associatedtype Connection:
  AblyChat.Connection`); the API Design Guidelines [1] touch on this,
  using IteratorProtocol as their example.

- Cons: It makes it more cumbersome to work with values whose type is
  that protocol (e.g. it's nicer for a user to work with `any Room` than
  `any RoomProtocol`).

I think that in our case it's much more likely that users are going to
want to use instances of these protocols than they are to wish to
implement these protocols (unlike, say, IteratorProtocol, which users
almost never consume directly), so I think that the cons outweight the
pros.

I haven't addressed any of the internal protocol naming here since it's
not required pre-v1.

[1] https://www.swift.org/documentation/api-design-guidelines/