Mark PaginatedResult as @MainActor
Not sure why I didn't do this in fc83fc1; probably I thought of it as a
data container more than a client, but really it's a client given its
methods for fetching further pages. (The compiler didn't require us to
make it @MainActor because our internal implementation is currently
going directly via the ably-cocoa paginated result type which isn't
marked as @MainActor, but we want to be sure that we will be able to
write a PaginatedResult implementation in any future client that _does_
enforce main actor isolation, e.g. upcoming ably-swift).