Remove dodgy MessageVersion Equatable implementation
This looks like it was a hack to get tests to pass.

The integration tests fix (areMessagesEqualModuloNonSerialVersionInfo)
just preserves the previous behaviour of the assertions. The ChatAPI
unit tests fix just asserts the values observed in the tests. I am not
familiar enough with the message versioning functionality to know if
this is the _right_ behaviour.