Remove Equatable conformance from all the options types
There's no reason for a user to wish to compare two of these, and we
want the flexibility to be able to add non-Equatable properties (e.g. a
callback) in the future.

TODO figure out how to get this to compile — we probably need to use
the options to create a separate internal Equatable type (a bit of a
nuisance)