Remove Semaphore library
I think that when we introduced this library we were initially making
heavier use of it; now we're just using it in the test suite.

Don't want our v1 release to have a dependency on a 0.x library; we'd
either need to pin the version and cause potential dependency version
clashes for other users (that is, if the author ever releases another
version), or not pin the version and have our library break if the
author releases a new 0.x version with breaking API changes.