Partially align Message.update signature with JS at d778026
Accept an OperationDetails instead of separate description and metadata,
and align the (currently-unused) UpdateMessageParams.

We can't fully align the signature yet (to not require a `Message` and
instead take a serial-like value and an UpdateMessageParams) because
we're waiting on updating `update` to use the server return value, which
is currently WIP in #392.