presence: remove event filter
In the rest of the SDK, we've not provided event filters on subscription
events - this is a fairly trivial thing for users to write (if/switch).
Also, when dealing with multiple events in the list, the if/switch is
required anyway and in many languages must be exhaustive, so the real
benefit is when you're listening to a single event (but still marginal).

Furthermore, the overloading is somewhat cumbersome to maintain.

This change removes the event filtering for presence.