Make Room's status API consistent with Connection
This undoes the public API change introduced in d51887c, in which we try
to make it clear from the type system whether a given room status is
guaranteed to have an error, guaranteed to have no error, or maybe has
an error.

I still think that the API from d51887c is the "correct" one, but I
suggested making this change in JS (see our #12) and nobody seemed
particularly keen to do so (particularly because the core SDKs, whose
errors we are largely just reflecting, don't expose such an API and
don't make these guarantees, leading us to have to guess what the
correct contract is).

So, change the API to match JS and our Connection type; that is, remove
the RoomStatus associated types and introduce `error` properties on Room
and RoomStatusChange. We can revisit this one day once we have a core
SDK API that makes these guarantees (e.g. in upcoming ably-swift).

Resolves #406.