Further naming tweaks
Change forSerial to withSerial in Messages.delete and Messages.update.
(In 103a812 I suggested forSerial because I thought withSerial would
look like the serial is being used as a parameter to change the nature
of the operation instead of identifying the message, but now that I've
seen "for serial" being used it just looks weird.)

And change MessageReactions delete(forMessageWithSerial:…) to
delete(fromMessageWithSerial:…); i.e. "delete reaction from the message
with this serial".