Throw our own ErrorInfo type instead of ARTErrorInfo
NSError (and users having to think about error domains in order to
interpret an error code) isn't really used in modern Swift libraries.
Instead, create an ErrorInfo struct and use that.

And do not expose the ErrorCode enum to the user; this makes us
consistent with our other SDKs, in which our website's documentation is
the authoritative source of information on the meaning of error codes,
and which avoids a two-tier system of error codes in which there's an
enum that describes the codes thrown internally by the SDK but doesn't
comprehensively cover those that the backend might return.

ErrorInfo documentation based on docstrings in spec repo but tweaked a
bit because they're not really fit for purpose at the moment (see [1]).

[1] https://github.com/ably/specification/issues/395