Align the error codes with the spec
Based on spec PR [1] at 23ca1db.

I've also added documentation for all of the InternalError cases to
describe what they represent. We can also use this documentation to
explain the associated data.

Note that this highlights that there are some removed spec points that
we're still implementing; things that we missed in e.g. the single
channel migration. We still need to address those but it's not in the
scope of the current task.

We still have some usage of the 40000 BadRequest code for errors not
specified by the spec, violating CHA-GP5; will revisit these errors
separately in #438.

Note that we no longer have any error codes with a variable status code;
will remove this mechanism in a separate commit in case we need to
restore it some time.

Resolves #426.

[1] https://github.com/ably/specification/pull/394