Don't use package manifest to treat warnings as errors
This reverts the configuration changes from a78e4ae. It turns out that
this is causing a compilation error of "error: conflicting options
'-warnings-as-errors' and '-suppress-warnings'" when trying to use Xcode
to build an app that uses the SDK.

Will create a minimal reproduction and raise a bug with Apple (have
created #441 for this and restoring a78e4ae once fixed).

I'm not sure why our example app wasn't catching this, but I guess there
must be a difference in the way that the SDK gets built when it's
included in the SDK via a workspace as opposed to via SPM. I've created
#442 for for making the example app build more realistic to try and
catch this sort of issue in the future; tried doing it as part of this
work but didn't find a good solution, and want to get this fix released.

Resolves #436.