example app: Improve experience on macOS and tvOS
It seems that the `flip` modifier that we were using to keep the scroll
view anchored to the bottom was causing issues with message reaction
buttons and the edit/delete button on these platforms. Switch to using
defaultScrollAnchor (Claude's suggestion; my SwiftUI knowledge is not
good enough to properly evaluate if this is the best approach) instead.
Seems to work well enough, though.