Fix hanging typing tests
I'm still noticing an occasional hang in the integration tests, but I
think that the majority of our test hangs were coming from the typing
tests, so I'm re-enabling the SPM test CI step in order to be sure that
these typing tests (which caused that CI step to hang) are fixed. Let's
keep an eye on the integration tests.

(It's good to have resolved most of the `swift test` hangs, because they
were also limiting the utility of Claude Code, which kept hanging when
trying to run the tests to verify its changes!)

I'm also still observing the iOS unit tests taking substantially
_longer_ to execute in CI than for other platforms. For example, the
result of one test run (with the integration tests disabled):

> ✔ Test run with 206 tests in 28 suites passed after 107.794 seconds.

where for other platforms the unit test execution time is below 1 second.

It looks like this is already covered by open issue #143, though.